Move TrackRepository into separate module
