+++ OTAnalytics/application/use_cases/highlight_intersections.py
@@ -9,7 +9,8 @@ from OTAnalytics.application.use_cases.section_repository import GetSectionsById
 from OTAnalytics.domain.event import EventRepository
 from OTAnalytics.domain.flow import FlowId, FlowRepository
 from OTAnalytics.domain.section import SectionId
-from OTAnalytics.domain.track import Track, TrackId, TrackIdProvider, TrackRepository
+from OTAnalytics.domain.track import Track, TrackId, TrackIdProvider
+from OTAnalytics.domain.track_repository import TrackRepository
 
 
 class IntersectionRepository(ABC):
