+++ OTAnalytics/application/use_cases/intersection_repository.py
@@ -8,7 +8,7 @@ from OTAnalytics.domain.section import (
     SectionListObserver,
     SectionRepositoryEvent,
 )
-from OTAnalytics.domain.track import TrackListObserver, TrackRepositoryEvent
+from OTAnalytics.domain.track_repository import TrackListObserver, TrackRepositoryEvent
 
 
 class ClearAllIntersections(SectionListObserver, TrackListObserver):
