+++ OTAnalytics/plugin_parser/otvision_parser.py
@@ -42,8 +42,8 @@ from OTAnalytics.domain.track import (
     TrackHasNoDetectionError,
     TrackId,
     TrackImage,
-    TrackRepository,
 )
+from OTAnalytics.domain.track_repository import TrackRepository
 from OTAnalytics.domain.video import PATH, SimpleVideo, Video, VideoReader
 from OTAnalytics.plugin_datastore.python_track_store import (
     PythonDetection,
