+++ tests/OTAnalytics/application/use_cases/test_highlight_intersections.py
@@ -30,13 +30,8 @@ from OTAnalytics.domain.event import Event, EventRepository
 from OTAnalytics.domain.filter import FilterElement
 from OTAnalytics.domain.flow import Flow, FlowId, FlowRepository
 from OTAnalytics.domain.section import Section, SectionId
-from OTAnalytics.domain.track import (
-    Detection,
-    Track,
-    TrackId,
-    TrackIdProvider,
-    TrackRepository,
-)
+from OTAnalytics.domain.track import Detection, Track, TrackId, TrackIdProvider
+from OTAnalytics.domain.track_repository import TrackRepository
 
 
 @pytest.fixture
