+++ tests/OTAnalytics/domain/test_track.py
@@ -3,11 +3,9 @@ from unittest.mock import MagicMock, Mock, call
 
 import pytest
 
-from OTAnalytics.domain.track import (
-    Track,
-    TrackDataset,
+from OTAnalytics.domain.track import Track, TrackDataset, TrackId
+from OTAnalytics.domain.track_repository import (
     TrackFileRepository,
-    TrackId,
     TrackListObserver,
     TrackObserver,
     TrackRepository,
